BVA/requirements.txt
accelerate
bitsandbytes
openai-whisper
pandas
prompt_toolkit
matplotlib==3.8.0
rich
scikit-learn==1.3.0
sentencepiece
torch
transformers
TTS==0.16.5
